Modificacion en servicios de Voluntarios

+++ Backend/resq/src/main/java/com/example/resq/Services/VoluntarioEmergenciaServices.java
@@ -10,34 +10,66 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/voluntario_emergencia")
+@RequestMapping("/api/voluntario_emergencias")
 public class VoluntarioEmergenciaServices {
     @Autowired
     VoluntarioEmergenciaRepository voluntarioEmergenciaRepository;
     @GetMapping
-    public List<VoluntarioEmergencia> findAllVoluntarioEmergencias(){
-        return voluntarioEmergenciaRepository.getAll();
+    public List<VoluntarioEmergencia> obtenerTodosVoluntariosEmergencia() {
+        List<VoluntarioEmergencia> voluntariosEmergencia = voluntarioEmergenciaRepository.getAll();
+        return  voluntariosEmergencia;
     }
     @GetMapping("/{id}")
-    public VoluntarioEmergencia findByIdVoluntarioEmergencia(@PathVariable("id") Integer id){
-        return voluntarioEmergenciaRepository.getById(id);
+    public VoluntarioEmergencia obtenerVoluntarioEmergenciaPorId(@PathVariable Integer id) {
+        VoluntarioEmergencia voluntarioEmergencia = voluntarioEmergenciaRepository.getById(id);
+        return voluntarioEmergencia;
     }
-    @DeleteMapping("/{id}")
-    public void deleteVoluntarioEmergencia(@PathVariable("id") Integer id){
-        voluntarioEmergenciaRepository.delete(id);
+    @GetMapping("/emergencia/{id}")
+    public Integer cantidadRegistrados(@PathVariable("id") Integer id) {
+        Integer salida = voluntarioEmergenciaRepository.cantidadVoluntarios(id);
+        return salida;
+    }
+    @GetMapping("/emergencia/{idEmergencia}/voluntario/{idVoluntario}")
+    public Boolean verificarRegistro(@PathVariable("idEmergencia") String emergencia,
+                                     @PathVariable("idVoluntario") String voluntario){
+        System.out.println("Verificando registro de " + voluntario + "en " + emergencia);
+        Integer idEmergencia = Integer.parseInt(emergencia);
+        Integer idVoluntario = Integer.parseInt(voluntario);
+        Boolean salida = voluntarioEmergenciaRepository.verificarRegistro(idEmergencia, idVoluntario);
+        return salida;
     }
     @PostMapping
-    public void saveVoluntarioEmergencia(@PathVariable("id") Integer id_emergencia,
-                                         @PathVariable("id") Integer id_usuario){
-        voluntarioEmergenciaRepository.save(id_emergencia, id_usuario);
+    public String crearVoluntarioEmergencia(@RequestBody VoluntarioEmergencia voluntarioEmergencia) {
+        System.out.println("id_emergencia: " + voluntarioEmergencia.getId_emergencia());
+        System.out.println("id_voluntario: " + voluntarioEmergencia.getId_voluntario());
+        voluntarioEmergenciaRepository.save(
+                voluntarioEmergencia.getId_emergencia(),
+                voluntarioEmergencia.getId_voluntario()
+        );
+        return "Relación creada correctamente";
     }
-    @GetMapping("/cantidad/{id_emergencia}")
-    public Integer cantidadVoluntarios(@PathVariable("id_emergencia") Integer id_emergencia){
-        return voluntarioEmergenciaRepository.cantidadVoluntarios(id_emergencia);
+    @PutMapping("/{id}")
+    public Boolean actualizarVoluntarioEmergencia(
+            @PathVariable Integer id,
+            @RequestBody VoluntarioEmergencia voluntarioEmergencia)
+    {
+        VoluntarioEmergencia voluntarioEmergenciaExistente = voluntarioEmergenciaRepository.getById(id);
+        if (voluntarioEmergenciaExistente != null) {
+            voluntarioEmergencia.setIdVoluntarioEmergencia(id);
+            voluntarioEmergenciaRepository.save(
+                    voluntarioEmergencia.getId_emergencia(),
+                    voluntarioEmergencia.getId_voluntario()
+            );
+            return true;
+        }
+        return false;
     }
-    @GetMapping("/verificar/{id_voluntario}/{id_emergencia}")
-    public Boolean verificarRegistro(@PathVariable("id_voluntario") Integer id_voluntario,
-                                     @PathVariable("id_emergencia") Integer id_emergencia){
-        return voluntarioEmergenciaRepository.verificarRegistro(id_voluntario, id_emergencia);
+
+    @DeleteMapping("/{id}")
+    public void eliminarVoluntarioEmergencia(@PathVariable Integer id) {
+        VoluntarioEmergencia voluntarioEmergencia = voluntarioEmergenciaRepository.getById(id);
+        if (voluntarioEmergencia != null) {
+            voluntarioEmergenciaRepository.delete(id);
+        }
     }
 }
